rhel85: split distro into multiple files
Split pipelines, inputs, and options into separate files for easier
navigation.
