+++ internal/distro/rhel85/stage_inputs.go
@@ -1,0 +1,51 @@
+package rhel85
+
+import (
+	osbuild "github.com/osbuild/osbuild-composer/internal/osbuild2"
+	"github.com/osbuild/osbuild-composer/internal/rpmmd"
+)
+
+func bootISOMonoStageInputs() *osbuild.BootISOMonoStageInputs {
+	rootfsInput := new(osbuild.BootISOMonoStageInput)
+	rootfsInput.Type = "org.osbuild.tree"
+	rootfsInput.Origin = "org.osbuild.pipeline"
+	rootfsInput.References = osbuild.BootISOMonoStageReferences{"name:anaconda-tree"}
+	return &osbuild.BootISOMonoStageInputs{
+		RootFS: rootfsInput,
+	}
+}
+
+func rpmStageInputs(specs []rpmmd.PackageSpec) *osbuild.RPMStageInputs {
+	stageInput := new(osbuild.RPMStageInput)
+	stageInput.Type = "org.osbuild.files"
+	stageInput.Origin = "org.osbuild.source"
+	stageInput.References = pkgRefs(specs)
+	return &osbuild.RPMStageInputs{Packages: stageInput}
+}
+
+func pkgRefs(specs []rpmmd.PackageSpec) osbuild.RPMStageReferences {
+	refs := make([]string, len(specs))
+	for idx, pkg := range specs {
+		refs[idx] = pkg.Checksum
+	}
+	return refs
+}
+
+func ostreePullStageInputs(origin, source, commitRef string) *osbuild.OSTreePullStageInputs {
+	pullStageInput := new(osbuild.OSTreePullStageInput)
+	pullStageInput.Type = "org.osbuild.ostree"
+	pullStageInput.Origin = origin
+
+	inputRefs := make(map[string]osbuild.OSTreePullStageReference)
+	inputRefs[source] = osbuild.OSTreePullStageReference{Ref: commitRef}
+	pullStageInput.References = inputRefs
+	return &osbuild.OSTreePullStageInputs{Commits: pullStageInput}
+}
+
+func xorrisofsStageInputs() *osbuild.XorrisofsStageInputs {
+	input := new(osbuild.XorrisofsStageInput)
+	input.Type = "org.osbuild.tree"
+	input.Origin = "org.osbuild.pipeline"
+	input.References = osbuild.XorrisofsStageReferences{"name:bootiso-tree"}
+	return &osbuild.XorrisofsStageInputs{Tree: input}
+}
